add docs for asset and cleanu
