+++ docs/source/classes/asset.rst
@@ -1,0 +1,7 @@
+Asset class
+===========
+
+.. autoclass:: starfish.asset.Asset
+    :members:
+    :undoc-members:
+    :show-inheritance:
